consensus: make `Header` blob fees u64 (#1377)
consensus: make Header blob fees u64

+++ crates/rpc-types-eth/src/block.rs
@@ -130,7 +130,7 @@ pub struct Header {
             with = "alloy_serde::quantity::opt"
         )
     )]
-    pub blob_gas_used: Option<u128>,
+    pub blob_gas_used: Option<u64>,
     /// Excess blob gas
     #[cfg_attr(
         feature = "serde",
@@ -140,7 +140,7 @@ pub struct Header {
             with = "alloy_serde::quantity::opt"
         )
     )]
-    pub excess_blob_gas: Option<u128>,
+    pub excess_blob_gas: Option<u64>,
     /// EIP-4788 parent beacon block root
     #[cfg_attr(feature = "serde", serde(default, skip_serializing_if = "Option::is_none"))]
     pub parent_beacon_block_root: Option<B256>,
@@ -170,7 +170,7 @@ impl Header {
     /// spec.
     ///
     /// Returns a `None` if no excess blob gas is set, no EIP-4844 support
-    pub fn next_block_excess_blob_gas(&self) -> Option<u128> {
+    pub fn next_block_excess_blob_gas(&self) -> Option<u64> {
         Some(calc_excess_blob_gas(self.excess_blob_gas?, self.blob_gas_used?))
     }
 }
